Add "inner" property to request errors
When using the sdk, it sometimes happen that the request executor fails
for reason that are not API related. This fix adds the original request
error to the wrappedError to help the code to respond to it.

The wrapped error now contains a property "inner" that has the original error