Fixing broken mobile navbar.
By adding in the required Javascript:

- bootstrap
- jquery

And modifying the navbar HTML elements slightly (convert the div to a nav
element, move the navbar inside of the container), I was able to get the menu
clickable.

There is still work to be done for styling -- the menu items are transparent, so
it looks bad.