Fixing the base URL option.
Previously, if you defined a custom base URL you would get an exception because the request library was being passed a `baseUrl` and a `uri`, both as fully-qualified URLs.  But it wants
the `uri` to be a relative path if the `baseUrl` is defined.

This commit fixes the problem, but always passing `uri` as a path, and doing a string replace on the passed `uri` to ensure that the base path is always set to the one that is configured by options.

IT tests are happy, small fixes for the usage of nock in our unit tests.